Text and talker information delineated, 3 separate nursery rhymes, good for computing analytics and word counts

Text and talker information delineated, 3 separate nursery rhymes, good for computing analytics and word counts

Usage

NurseryRhymes

Format

## "NurseryRhymes" A data.frame with 100 observations, 2 vars:

Event_ID

factor 3 different simulated conversations

Participant_ID

fictional speaker names, 2 each conversation

Text_Raw

simulated language production, actually looped phrases from nursery rhymes

...