Text and talker information delineated, 3 separate nursery rhymes, good for computing analytics and word counts

Text and talker information delineated, 3 separate nursery rhymes, good for computing analytics and word counts

Usage

NurseryRhymes_Prepped

Format

## "NurseryRhymes_Prepped" A data.frame with 1507 x 7 observations, 5 vars:

Event_ID

factor 3 different simulated conversations

Participant_ID

fictional speaker names, 2 each conversation

Exchange_Count

sequential numbering of exchanges by conversation, 1 exchange = 2 turns

Turn_Count

sequential numbering of turns by conversation

Text_Clean

content words

emo_anger

raw value of anger salience yoked to each word

...