summarize_dyads_covar

Calculates Spearman rank correlation and lagged Pearson correlation between interlocutor time series for each dimension and transcript.

Usage

summarize_dyads_covar(aligned_ts_df, lags = c(-3, -2, -1, 0, 1, 2, 3))

Arguments

aligned_ts_df

Dataframe produced in the align_dyads function

lags

A vector of signed integers specifying lags that should be included for Pearson correlation. Negative integers are used for leads.

Value

A data frame containing the Spearman and Pearson correlation for each transcript. Results are organized as column and data is pivoted longer by dimension for readability.